minor fix
2005/01/13 03:31:25-05:00 (none)!donp
Minor fix.

2005/01/13 03:15:30-05:00 (none)!donp
Change mode to -rw-rw-rw-

2005/01/13 03:15:05-05:00 (none)!donp
Change mode to -rw-rw-r--

2005/01/13 03:11:24-05:00 (none)!donp
Fix to changes made in libsefs.

2005/01/07 15:21:29-05:00 tresys.com!bwhalen
Change searchcon api to work with new libsefs setup that allows for multiple search types

2005/01/03 14:30:16-05:00 tresys.com!bwhalen
Had to move define of fsdata.h to top of defines.

2004/12/30 14:59:47-05:00 tresys.com!bwhalen
Switched to newer libsefs api

2004/12/23 10:49:12-05:00 tresys.com!bwhalen
Edited searchcon to work with the new api.

2004/12/21 14:56:29-05:00 tresys.com!bwhalen
put in a check to see if they just want all known types.

2004/12/21 13:29:27-05:00 tresys.com!bwhalen
Put to use the new datastructs from libsefs that will allow us to search and
print.

2004/12/20 16:19:52-05:00 tresys.com!bwhalen
Edited searchon so that it uses the new api

(Logical change 1.514)