fix merge problems
2005/01/13 04:20:22-05:00 (none)!donp
Fix to return object classes

2005/01/13 03:11:23-05:00 (none)!donp
Changed to const strings.

2005/01/11 13:46:55-05:00 homenetwork!brandon
Indexing on wrong data for object classes.

2005/01/11 11:26:07-05:00 homenetwork!brandon
Fixed the indexes so we have faster search times,
and changed our statement populate to correctly create the select stmt

2005/01/07 15:21:29-05:00 tresys.com!bwhalen
Move sqlite3.h to top

2005/01/06 13:18:30-05:00 tresys.com!bwhalen
Auto merged

2005/01/06 13:11:00-05:00 tresys.com!bwhalen
Remove debugging statements

2005/01/06 12:58:20-05:00 (none)!method
check

2005/01/06 11:09:35-05:00 tresys.com!bwhalen
More memory leak work fixes

2005/01/06 10:33:17-05:00 tresys.com!bwhalen
Put in changes to handle found leak, and started setting up structs for
mulitple types/users

2005/01/05 13:57:25-05:00 tresys.com!bwhalen
now use the sqlite header file in our directory tree

2005/01/03 14:30:16-05:00 tresys.com!bwhalen
Had to move around defines to allow for stat64

2004/12/30 14:59:46-05:00 tresys.com!bwhalen
Moved some defines, changed the api of print_valid_object classes

2004/12/30 10:23:15-05:00 tresys.com!bwhalen
Moved over headers from the .h file that are no longer needed as part of the
api.  Also converted to stat64 to handle large files.

2004/12/29 14:42:25-05:00 tresys.com!bwhalen
Put in a function to close an open db, handle destroying fsdata on an error.

2004/12/23 10:49:12-05:00 tresys.com!bwhalen
-Put in a callback to handle recieving the result of a select count(*)
-Changed the search callback so that we now keep a global pointer to the last
entry in the linked list to try to speed up returning large result sets, we
now have a good many global variables(and having gone to UMD where globals
are bad) I am debating creating a struct that has all these data types and
passing that to the callbacks, the reason I don't do that now is callbacks can
only accept 1 pointer.
-Fixed an issue I had with finding the correct object class.  Object class is
defined at 1,2,4,8...etc but we have this function that finds an objects class
location in an array so I was getting back incorrect data.  I created a function
so I can pass in a string object class and get back the correct defined int.
-Added in a function that will return the known things(types,users,
object classes, and paths) this will be useful for the interface when we make
it.
-Added in fcns to print/destroy the returned char** array returned by the
known search fcn.
-Reordered the search string building(in the search fcn) so that now its
in logical order you can check to make sure all cases are covered, and easier
to find a problem if you have one.
-I took out some debuggging statements, and also added in some comments.
-Also I think I fixed an issue with searching for paths that has to do with a
filename have "" in it.  Using an sqlite3 fcn that will double any ' and
switched from using "" to qoute my strings to '' hoping that now I won't run
into errors.  From my testing this works.

2004/12/21 14:56:29-05:00 tresys.com!bwhalen
Added ability to return all known types, changed how we delete libsefs created
structs.

2004/12/21 13:29:27-05:00 tresys.com!bwhalen
Have now changed it so when user is not using regular expressions we use
sqlite built in compare on the strings.  Also added into callback fcn the
loading of the return data struct.  Also added in a destroy and print fcn on
that data struct.

2004/12/20 16:19:52-05:00 tresys.com!bwhalen
Edited some of the old returns to better handle memory, created the search
functions so that now we can search the database.  This is mostly done
except for returning something that can be used by a calling routine.

2004/12/17 16:19:33-05:00 tresys.com!bwhalen
put in search keys function

2004/12/17 14:35:11-05:00 tresys.com!bwhalen
moved over old data structs and api from .h to hide it so that we have a smaller and
cleaner api.

2004/12/17 10:45:47-05:00 tresys.com!bwhalen
merge in sqlite to allow us to save file as a sqlite3 db.

(Logical change 1.514)