Make the SWIG files compatible with SWIG 3
SWIG changed the way types are declared, so that this code:

    typedef struct my_struct {} my_struct_t;
    %extend my_struct_t {
        void my_fun(int arg) {
            my_struct_my_fun(self, arg); /* code in a .so library */
        }
    }

... generated a wrapper function named my_struct_t_my_fun which calls
my_struct_my_fun.  In SWIG 3 (and maybe some SWIG 2 versions), this
wrapper is now named my_struct_my_fun, which conflicts with the
library-provided my_struct_my_fun, thus breaking the build process.

To fix this, rename "my_fun" to "wrap_my_fun" and use SWIG %rename to
keep the name in the wrapper which is used in Python/Tcl/Java:

        %rename(my_fun) wrap_my_fun;
        void wrap_my_fun(int arg) {
            my_struct_my_fun(self, arg);
        }

This idea comes from a Gentoo patch which has been written to address
https://bugs.gentoo.org/show_bug.cgi?id=430262 and is no longer used
(see comment 10 of Gentoo bug 430262 for more details).

Tested with SWIG 3.0.5 and 2.0.12, with SELinux userspace tools & libraries 2.4.

Relevant SWIG documentation page about %rename:
http://www.swig.org/Doc3.0/SWIGPlus.html#SWIGPlus_ambiguity_resolution_renaming